fix(ФормаОсновная): fix: add pattern command for repl
repl options move to class Player
User can interacts with this, but don't know how it's worked
